+++ backend/.adminjs/entry.js
@@ -3,3 +3,5 @@ import Dashboard from '../src/admin/components/dashboard'
 AdminJS.UserComponents.Dashboard = Dashboard
 import JsonShow from '../src/admin/components/json-show'
 AdminJS.UserComponents.JsonShow = JsonShow
+import RestoreBackup from '../src/admin/components/restore-backup'
+AdminJS.UserComponents.RestoreBackup = RestoreBackup